+++ src/standard_harness/docsops/__init__.py
@@ -1,0 +1,2 @@
+"""Documentation operations helpers."""
+